refactor: extract backend and CLI code into separate crates
Create devenv-core crate containing shared backend functionality (CLI parsing,
config, Nix args, and Nix backend) and devenv-snix-backend crate for SNIX
backend implementation. Update workspace and crate dependencies accordingly.
