+++ devenv-core/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "devenv-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[features]
+snix = []
+
+[dependencies]
+devenv-eval-cache.workspace = true
+nix-conf-parser.workspace = true
+pathdiff.workspace = true
+
+async-trait.workspace = true
+clap.workspace = true
+miette.workspace = true
+once_cell.workspace = true
+schemars.workspace = true
+schematic.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+serde_yaml.workspace = true
+ser_nix.workspace = true
+tokio.workspace = true
+tracing.workspace = true
