+++ devenv-core/src/config.rs
@@ -868,373 +868,4 @@ mod tests {
             "Default config should serialize to empty YAML"
         );
     }
-
-    #[test]
-    fn test_devenv_yaml_import_merging() {
-        let fixture_path = Path::new("fixtures/config/import-merging");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        assert!(config.allow_unfree);
-        assert!(config.allow_broken);
-        assert!(config.impure);
-        assert_eq!(config.inputs.len(), 2);
-        assert_eq!(
-            config.inputs["nixpkgs"].url,
-            Some("github:NixOS/nixpkgs/nixos-23.11".to_string())
-        );
-        assert_eq!(
-            config.inputs["flake-utils"].url,
-            Some("github:numtide/flake-utils".to_string())
-        );
-    }
-
-    #[test]
-    fn test_relative_import_paths() {
-        let fixture_path = Path::new("fixtures/config/relative-imports");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        assert!(config.allow_unfree);
-        assert!(config.allow_broken);
-    }
-
-    #[test]
-    fn test_circular_import_prevention() {
-        let fixture_path = Path::new("fixtures/config/circular-imports");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        assert!(config.allow_unfree);
-        assert!(config.allow_broken);
-    }
-
-    #[test]
-    fn test_nixpkgs_config_merging() {
-        let fixture_path = Path::new("fixtures/config/nixpkgs-merging");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        let nixpkgs = config.nixpkgs.expect("nixpkgs should be present");
-        assert!(nixpkgs.config_.allow_unfree);
-        assert!(nixpkgs.config_.allow_broken);
-        assert!(nixpkgs.config_.cuda_support);
-        assert_eq!(nixpkgs.config_.cuda_capabilities.len(), 3);
-        assert!(
-            nixpkgs
-                .config_
-                .cuda_capabilities
-                .contains(&"7.5".to_string())
-        );
-        assert!(
-            nixpkgs
-                .config_
-                .cuda_capabilities
-                .contains(&"8.0".to_string())
-        );
-        assert!(
-            nixpkgs
-                .config_
-                .cuda_capabilities
-                .contains(&"8.6".to_string())
-        );
-        assert_eq!(nixpkgs.config_.permitted_insecure_packages.len(), 1);
-    }
-
-    #[test]
-    fn test_nested_imports() {
-        let fixture_path = Path::new("fixtures/config/nested-imports");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        assert!(config.allow_unfree);
-        assert!(config.allow_broken);
-        assert!(config.impure);
-        assert_eq!(config.inputs.len(), 1);
-        assert_eq!(
-            config.inputs["custom"].url,
-            Some("github:user/repo".to_string())
-        );
-    }
-
-    #[test]
-    fn test_duplicate_imports() {
-        let fixture_path = Path::new("fixtures/config/duplicate-imports");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        // Test that shared is imported multiple times but allowUnfree is still true
-        assert!(config.allow_unfree);
-
-        // Test array appending with duplicates
-        // Should contain all entries including duplicates from append_vec merge strategy
-        assert!(
-            config
-                .permitted_insecure_packages
-                .contains(&"openssl-1.0.2".to_string())
-        );
-        assert!(
-            config
-                .permitted_insecure_packages
-                .contains(&"shared-package".to_string())
-        );
-        assert!(
-            config
-                .permitted_insecure_packages
-                .contains(&"module1-package".to_string())
-        );
-        assert!(
-            config
-                .permitted_insecure_packages
-                .contains(&"module2-package".to_string())
-        );
-
-        // Check that duplicates are preserved (append_vec doesn't deduplicate)
-        let openssl_count = config
-            .permitted_insecure_packages
-            .iter()
-            .filter(|&p| p == "openssl-1.0.2")
-            .count();
-        assert!(
-            openssl_count >= 2,
-            "Expected duplicate openssl entries, got {}",
-            openssl_count
-        );
-    }
-
-    #[test]
-    fn test_boolean_merging_last_wins() {
-        let fixture_path = Path::new("fixtures/config/boolean-merging");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        // With replace strategy, last import wins
-        // Order: base (false) -> first (true) -> second (false) -> third (true for allowUnfree only)
-        assert!(config.allow_unfree); // third sets this to true
-        assert!(!config.allow_broken); // second set to false, third doesn't change
-        assert!(!config.impure); // second set to false, third doesn't change
-    }
-
-    #[test]
-    fn test_array_duplicates_behavior() {
-        let fixture_path = Path::new("fixtures/config/array-duplicates");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        let nixpkgs = config.nixpkgs.expect("nixpkgs should be present");
-
-        // With append_vec, all entries are preserved including duplicates
-        // Base: ["7.5", "7.5", "8.0"] + sub1: ["8.0", "8.6", "8.0"] + sub2: ["7.5", "9.0"]
-        let capabilities = &nixpkgs.config_.cuda_capabilities;
-        assert!(capabilities.contains(&"7.5".to_string()));
-        assert!(capabilities.contains(&"8.0".to_string()));
-        assert!(capabilities.contains(&"8.6".to_string()));
-        assert!(capabilities.contains(&"9.0".to_string()));
-
-        // Count duplicates
-        let count_7_5 = capabilities.iter().filter(|&c| c == "7.5").count();
-        let count_8_0 = capabilities.iter().filter(|&c| c == "8.0").count();
-        assert!(
-            count_7_5 >= 3,
-            "Expected at least 3 occurrences of 7.5, got {}",
-            count_7_5
-        );
-        assert!(
-            count_8_0 >= 3,
-            "Expected at least 3 occurrences of 8.0, got {}",
-            count_8_0
-        );
-    }
-
-    #[test]
-    fn test_complex_nixpkgs_merging() {
-        let fixture_path = Path::new("fixtures/config/complex-nixpkgs");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        let nixpkgs = config.nixpkgs.expect("nixpkgs should be present");
-
-        // Top-level nixpkgs config
-        assert!(nixpkgs.config_.allow_unfree); // from base
-        assert!(nixpkgs.config_.allow_broken); // from platform-specific
-        assert!(nixpkgs.config_.cuda_support); // from cuda-config
-
-        // Per-platform configs (merge_btreemap replaces entire entries)
-        let x86_linux = &nixpkgs.per_platform["x86_64-linux"];
-        // The entire x86_64-linux entry is replaced by cuda-config which only has cudaCapabilities
-        assert!(!x86_linux.cuda_support); // default false, as cuda-config replaces the whole entry
-        assert_eq!(x86_linux.cuda_capabilities, vec!["9.0"]); // from cuda-config (last wins)
-
-        let x86_darwin = &nixpkgs.per_platform["x86_64-darwin"];
-        assert!(x86_darwin.allow_unfree); // from platform-specific
-
-        let aarch64_darwin = &nixpkgs.per_platform["aarch64-darwin"];
-        assert!(aarch64_darwin.allow_broken); // from platform-specific
-        assert!(!aarch64_darwin.cuda_support); // default false (not merged from base)
-    }
-
-    #[test]
-    fn test_path_edge_cases() {
-        let fixture_path = Path::new("fixtures/config/path-edge-cases");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        // Should load normal and "with-spaces in name" but silently ignore non-existent
-        assert!(config.allow_unfree); // from normal
-        assert!(config.allow_broken); // from "with-spaces in name"
-    }
-
-    #[test]
-    fn test_empty_configs() {
-        let fixture_path = Path::new("fixtures/config/empty-configs");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        // Empty base config with imports should still work
-        assert!(config.allow_unfree); // from with-content
-
-        // Other fields should have default values
-        assert!(!config.allow_broken);
-        assert!(!config.impure);
-        assert!(config.imports.len() >= 3); // base imports are preserved
-    }
-
-    #[test]
-    fn test_import_order_diamond_pattern() {
-        let fixture_path = Path::new("fixtures/config/import-order");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        // All imports should be processed despite diamond pattern
-        assert!(config.allow_unfree); // from b
-        assert!(config.allow_broken); // from c
-        assert!(config.impure); // from d (imported by both b and c)
-
-        // Input from d should only appear once despite being imported twice
-        assert_eq!(config.inputs.len(), 1);
-        assert_eq!(
-            config.inputs["shared-input"].url,
-            Some("github:shared/repo".to_string())
-        );
-    }
-
-    #[test]
-    fn test_optional_configs_merging() {
-        let fixture_path = Path::new("fixtures/config/optional-configs");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        // Clean config should be overridden (replace strategy for nested Option)
-        let clean = config.clean.expect("clean should be present");
-        assert!(!clean.enabled); // overridden from true to false
-        assert_eq!(clean.keep, vec!["build", "dist"]); // completely replaced
-
-        // Secretspec should be added
-        let secretspec = config.secretspec.expect("secretspec should be present");
-        assert!(secretspec.enable);
-        assert_eq!(secretspec.profile, Some("dev".to_string()));
-        assert_eq!(secretspec.provider, Some("aws".to_string()));
-    }
-
-    #[test]
-    fn test_path_traversal_prevention() {
-        let fixture_path = Path::new("fixtures/config/path-traversal");
-        let result = Config::load_from(fixture_path);
-
-        assert!(result.is_err(), "Expected error but got: {:?}", result);
-        let error_message = result.unwrap_err().to_string();
-        // Check for path traversal error (message varies depending on git repo detection)
-        assert!(
-            error_message.contains("resolves outside the")
-                && error_message.contains("which is not allowed"),
-            "Expected path traversal error, got: {}",
-            error_message
-        );
-    }
-
-    #[test]
-    fn test_missing_import_file_error() {
-        let fixture_path = Path::new("fixtures/config/empty-configs");
-        // This test works because empty-configs has imports to non-existent directories
-        // which are silently ignored (no devenv.yaml file exists)
-        let config = Config::load_from(fixture_path)
-            .expect("Missing import files should be silently ignored");
-        assert!(config.allow_unfree); // Should still load successfully
-    }
-
-    #[test]
-    fn test_input_url_normalization() {
-        let fixture_path = Path::new("fixtures/config/input-url-normalization");
-        let config = Config::load_from(fixture_path).expect("Failed to load config");
-
-        // Input URLs from nested configs should be normalized to be relative to base path
-        assert_eq!(config.inputs.len(), 2);
-
-        // url: ../ from subdir should become ./ from base (canonicalized to current dir)
-        let root_url = config.inputs["root"]
-            .url
-            .as_ref()
-            .expect("root url should exist");
-        // Canonicalized paths may be "." or an empty relative path
-        assert!(
-            root_url == "./" || root_url == "./." || root_url == ".",
-            "Expected normalized path but got: {}",
-            root_url
-        );
-        assert!(!config.inputs["root"].flake);
-
-        // url: path:../../other from subdir - non-existent path won't be canonicalized
-        // diff_paths preserves the relative structure: subdir/../../other
-        assert_eq!(
-            config.inputs["other"].url,
-            Some("path:subdir/../../other".to_string())
-        );
-        assert!(!config.inputs["other"].flake);
-
-        // Imports should be normalized too
-        assert_eq!(config.imports.len(), 1);
-        assert_eq!(config.imports[0], "./subdir");
-    }
-
-    #[test]
-    fn test_compose_example() {
-        // Test the examples/compose case where projectB has an input with relative URL
-        let fixture_path = Path::new("../examples/compose");
-        if !fixture_path.exists() {
-            // Skip if examples don't exist (not in devenv directory)
-            return;
-        }
-
-        let config = Config::load_from(fixture_path).expect("Failed to load compose config");
-
-        // Check imports include both file-based and input-based imports
-        assert_eq!(
-            config.imports.len(),
-            3,
-            "Should have 3 imports (2 file + 1 input-based)"
-        );
-        assert!(config.imports.contains(&"./projectA".to_string()));
-        assert!(config.imports.contains(&"./projectB".to_string()));
-        assert!(
-            config.imports.contains(&"root/projectA".to_string()),
-            "Should contain input-based import from nested config"
-        );
-
-        // Check root input from projectB is merged and normalized
-        if let Some(root_input) = config.inputs.get("root") {
-            println!("Root input URL: {:?}", root_input.url);
-            // URL should be normalized from projectB's ../ to ./
-            let url = root_input.url.as_ref().expect("root should have URL");
-            assert!(
-                !url.contains("../"),
-                "URL should be normalized, got: {}",
-                url
-            );
-        }
-    }
-
-    #[test]
-    fn test_invalid_yaml_error_context() {
-        use std::fs;
-        let temp_dir = tempfile::tempdir().expect("Failed to create temp dir");
-        let invalid_yaml_path = temp_dir.path().join("devenv.yaml");
-
-        fs::write(&invalid_yaml_path, "invalid: yaml: content: [").expect("Failed to write file");
-
-        let result = Config::load_from(temp_dir.path());
-        assert!(result.is_err());
-        let error_message = result.unwrap_err().to_string();
-        assert!(
-            error_message.contains("devenv.yaml") || error_message.contains("Failed to load"),
-            "Error should mention the file that failed to load: {}",
-            error_message
-        );
-    }
 }
