+++ devenv/src/lib.rs
@@ -1,17 +1,19 @@
-pub mod cachix;
 pub mod cli;
-pub mod config;
 mod devenv;
 pub mod log;
 pub mod mcp;
 pub(crate) mod nix;
-pub mod nix_args;
-pub mod nix_backend;
 pub mod nix_log_bridge;
-#[cfg(feature = "snix")]
-pub(crate) mod snix_backend;
 mod util;
 
-pub use cli::{GlobalOptions, default_system};
+#[cfg(feature = "snix")]
+pub use devenv_snix_backend;
+
 pub use devenv::{DIRENVRC, DIRENVRC_VERSION, Devenv, DevenvOptions, ProcessOptions};
 pub use devenv_tasks as tasks;
+
+// Re-export core types from devenv-core for convenience
+pub use devenv_core::{
+    CachixCacheInfo, CachixManager, CachixPaths, Config, DevenvPaths, GlobalOptions, NixArgs,
+    NixBackend, Options, SecretspecData, default_system,
+};
